requirements_test.txt
-r requirements_dev.txt
pytest-homeassistant-custom-component==0.11.6
